Updating app requirements for Django 1.8 support.
The Django version in the requirements has been upgraded to version 1.8.
django-cms>3.1.1 is now required for Django 1.8 support. A fork of
django-shop is also required for Django 1.8 support. django-shop was
removed from the install_requires in setup.py because it was installing
a clashing version on top of the requirement specified in the
requirements/common.txt. This would cause South to also be installed,
which can cause problems in some Django 1.8 and newer projects.
The test settings have also been updated to reflect this upgrade.