Install django-shop with an egg of "django-shop" not "shop"
Otherwise, pip might be confused into pulling in an older version of
"django-shop" not realizing that the thing we pulled in as "shop" is
already a git-master of the same thing.

And since old versions of django-shop, in turn, pull in dependencies
(such as South) that break other packages, we really need to avoid
that.